backend/requirements.txt
Django>=4.2,<5.3
djangorestframework>=3.15,<3.17
django-filter>=24.2,<25.0
psycopg[binary]>=3.1,<3.3
celery>=5.4,<5.6
redis>=5.0,<6.0
django-celery-results>=2.5,<2.7
python-dotenv>=1.0,<2.0
gunicorn>=22.0,<23.0
